Plastics: approximate Scala Product instances and more
