+++ app/core/src/main/scala/com/qwery/runtime/datatypes/NumericType.scala
@@ -2,7 +2,7 @@ package com.qwery.runtime.datatypes
 
 import com.qwery.language.dieUnsupportedConversion
 import com.qwery.language.models.{ColumnType, Expression}
-import com.qwery.runtime.RuntimeClass.implicits.RuntimeClassConstructorSugar
+import com.qwery.runtime.plastics.RuntimeClass.implicits.RuntimeClassConstructorSugar
 import com.qwery.runtime.{Boolean2Int, Scope}
 import com.qwery.runtime.instructions.expressions.NamedFunctionCall
