+++ app/core/src/main/scala/com/qwery/runtime/datatypes/StringType.scala
@@ -2,7 +2,7 @@ package com.qwery.runtime.datatypes
 
 import com.qwery.implicits.MagicImplicits
 import com.qwery.language.models.{ColumnType, Instruction}
-import com.qwery.runtime.RuntimeClass.implicits.RuntimeClassConstructorSugar
+import com.qwery.runtime.plastics.RuntimeClass.implicits.RuntimeClassConstructorSugar
 import com.qwery.runtime.devices.{FieldMetadata, QMap, RowCollection}
 import com.qwery.runtime.{INT_BYTES, Scope}
 import com.qwery.util.ByteBufferHelper.{DataTypeBuffer, DataTypeByteBuffer}
