+++ app/core/src/main/scala/com/qwery/runtime/instructions/expressions/Infix.scala
@@ -4,7 +4,7 @@ import com.qwery.language.models.Expression.implicits.RichAliasable
 import com.qwery.language.models._
 import com.qwery.language.{ExpressionChainParser, HelpDoc, SQLCompiler, TokenStream}
 import com.qwery.runtime.QweryVM.implicits.RichScalaAny
-import com.qwery.runtime.RuntimeClass.implicits.{RuntimeClassExpressionSugar, RuntimeClassInstanceSugar}
+import com.qwery.runtime.plastics.RuntimeClass.implicits.{RuntimeClassExpressionSugar, RuntimeClassInstanceSugar}
 import com.qwery.runtime.devices.QMap
 import com.qwery.runtime.instructions.expressions.RuntimeExpression.RichExpression
 import com.qwery.runtime.{QweryVM, Scope}
