+++ app/core/src/main/scala/com/qwery/runtime/instructions/invocables/DefineImplicit.scala
@@ -4,7 +4,8 @@ import com.qwery.language.HelpDoc.{CATEGORY_SYSTEMS, PARADIGM_OBJECT_ORIENTED}
 import com.qwery.language.models.{Atom, CodeBlock, Instruction}
 import com.qwery.language.{HelpDoc, InvokableParser, SQLCompiler, SQLTemplateParams, TokenStream}
 import com.qwery.runtime.instructions.functions.NamedFunction
-import com.qwery.runtime.{RuntimeClass, Scope, VirtualMethod}
+import com.qwery.runtime.plastics.{RuntimeClass, VirtualMethod}
+import com.qwery.runtime.Scope
 import com.qwery.util.LogUtil
 import qwery.io.IOCost
