+++ app/core/src/main/scala/com/qwery/runtime/instructions/invocables/Import.scala
@@ -5,7 +5,8 @@ import com.qwery.language.models.Expression
 import com.qwery.language.{HelpDoc, InvokableParser, SQLCompiler, SQLTemplateParams, TokenStream}
 import com.qwery.runtime.datatypes.AnyType
 import com.qwery.runtime.instructions.expressions.RuntimeExpression.RichExpression
-import com.qwery.runtime.{QweryVM, RuntimeClass, Scope}
+import com.qwery.runtime.plastics.RuntimeClass
+import com.qwery.runtime.{QweryVM, Scope}
 import com.qwery.util.OptionHelper.OptionEnrichment
 import qwery.io.IOCost
