+++ app/core/src/main/scala/com/qwery/runtime/instructions/invocables/ImportImplicitClass.scala
@@ -4,7 +4,8 @@ import com.qwery.language.HelpDoc.{CATEGORY_SYSTEMS, PARADIGM_OBJECT_ORIENTED}
 import com.qwery.language.models.Expression
 import com.qwery.language.{HelpDoc, InvokableParser, SQLCompiler, SQLTemplateParams, TokenStream}
 import com.qwery.runtime.instructions.expressions.RuntimeExpression.RichExpression
-import com.qwery.runtime.{RuntimeClass, Scope}
+import com.qwery.runtime.{Scope}
+import com.qwery.runtime.plastics.RuntimeClass
 import com.qwery.util.OptionHelper.OptionEnrichment
 import qwery.io.IOCost
