+++ app/core/src/main/scala/com/qwery/runtime/instructions/jvm/InvokeVirtualMethod.scala
@@ -3,7 +3,7 @@ package com.qwery.runtime.instructions.jvm
 import com.qwery.language.HelpDoc.{CATEGORY_REFLECTION, PARADIGM_OBJECT_ORIENTED}
 import com.qwery.language.models.Expression
 import com.qwery.language.{ExpressionChainParser, HelpDoc, SQLCompiler, TokenStream}
-import com.qwery.runtime.RuntimeClass.implicits.RuntimeClassInstanceSugar
+import com.qwery.runtime.plastics.RuntimeClass.implicits.RuntimeClassInstanceSugar
 import com.qwery.runtime.instructions.expressions.{NamedFunctionCall, RuntimeExpression}
 import com.qwery.runtime.{QweryVM, Scope}
 import com.qwery.util.OptionHelper.OptionEnrichment
