+++ app/core/src/main/scala/com/qwery/runtime/instructions/jvm/MembersOf.scala
@@ -2,9 +2,9 @@ package com.qwery.runtime.instructions.jvm
 
 import com.qwery.language.HelpDoc.{CATEGORY_REFLECTION, PARADIGM_FUNCTIONAL}
 import com.qwery.language.models.{Expression, ParameterLike}
-import com.qwery.runtime.Plastic.implicits.MethodNameConverter
-import com.qwery.runtime.RuntimeClass.decodeModifiers
-import com.qwery.runtime.RuntimeClass.implicits.RuntimeClassConstructorSugar
+import com.qwery.runtime.plastics.Plastic.implicits.MethodNameConverter
+import com.qwery.runtime.plastics.RuntimeClass.decodeModifiers
+import com.qwery.runtime.plastics.RuntimeClass.implicits.RuntimeClassConstructorSugar
 import com.qwery.runtime.datatypes.{AnyType, DataType, StringType, TableType}
 import com.qwery.runtime.devices.RecordCollectionZoo.MapToRow
 import com.qwery.runtime.devices.RowCollectionZoo._
@@ -13,7 +13,8 @@ import com.qwery.runtime.instructions.expressions.TableExpression
 import com.qwery.runtime.instructions.functions.{FunctionCallParserE1, ScalarFunctionCall}
 import com.qwery.runtime.instructions.queryables.RuntimeQueryable
 import com.qwery.runtime.instructions.jvm.MembersOf.ParameterLikeToScalaCode
-import com.qwery.runtime.{QweryVM, RuntimeClass, Scope}
+import com.qwery.runtime.plastics.RuntimeClass
+import com.qwery.runtime.{QweryVM, Scope}
 import com.qwery.util.OptionHelper.OptionEnrichment
 import qwery.io.IOCost
