+++ app/core/src/main/scala/com/qwery/runtime/instructions/jvm/ObjectOf.scala
@@ -2,7 +2,7 @@ package com.qwery.runtime.instructions.jvm
 
 import com.qwery.language.HelpDoc.{CATEGORY_REFLECTION, PARADIGM_OBJECT_ORIENTED}
 import com.qwery.language.models.Expression
-import com.qwery.runtime.RuntimeClass.getObjectByName
+import com.qwery.runtime.plastics.RuntimeClass.getObjectByName
 import com.qwery.runtime.instructions.expressions.RuntimeExpression
 import com.qwery.runtime.instructions.expressions.RuntimeExpression.RichExpression
 import com.qwery.runtime.instructions.functions.{FunctionCallParserE1, ScalarFunctionCall}
