+++ app/core/src/main/scala/com/qwery/util/StringRenderHelper.scala
@@ -2,7 +2,7 @@ package com.qwery.util
 
 import com.qwery.implicits.MagicImplicits
 import com.qwery.language.models.Instruction
-import com.qwery.runtime.Plastic.tupleToSeq
+import com.qwery.runtime.plastics.Tuples.tupleToSeq
 import com.qwery.runtime.datatypes.{DataType, StringType}
 import com.qwery.runtime.devices.{QMap, Row, RowCollection}
 import com.qwery.util.JVMSupport.NormalizeAny
